Add MiniMessage

+++ Insights-API/src/main/java/dev/frankheijden/insights/api/config/notifications/ProgressNotification.java
@@ -2,6 +2,6 @@ package dev.frankheijden.insights.api.config.notifications;
 
 public interface ProgressNotification extends Notification {
 
-    ProgressNotification progress(double progress);
+    ProgressNotification progress(float progress);
 
 }
